+++ dataforge-meta/src/commonMain/kotlin/space/kscience/dataforge/meta/Configurable.kt
@@ -5,21 +5,21 @@ import space.kscience.dataforge.names.Name
 import kotlin.properties.ReadWriteProperty
 
 /**
- * A container that holds a [Config].
+ * A container that holds a [ObservableMeta].
  */
 public interface Configurable {
     /**
      * Backing config
      */
-    public val config: Config
+    public val config: ObservableMeta
 }
 
 
 public fun <T : Configurable> T.configure(meta: Meta): T = this.apply { config.update(meta) }
 
 @DFBuilder
-public inline fun <T : Configurable> T.configure(action: Config.() -> Unit): T = apply { config.apply(action) }
+public inline fun <T : Configurable> T.configure(action: ObservableMeta.() -> Unit): T = apply { config.apply(action) }
 
 /* Node delegates */
 
-public fun Configurable.config(key: Name? = null): ReadWriteProperty<Any?, Config?> = config.node(key)
+public fun Configurable.config(key: Name? = null): ReadWriteProperty<Any?, ObservableMeta?> = config.node(key)