+++ dataforge-meta/src/commonTest/kotlin/space/kscience/dataforge/meta/ConfigTest.kt
@@ -6,7 +6,7 @@ import kotlin.test.assertEquals
 class ConfigTest {
     @Test
     fun testIndexedWrite(){
-        val config = Config()
+        val config = MetaBuilder()
         config["a[1].b"] = 1
         assertEquals(null, config["a.b"].int)
         assertEquals(1, config["a[1].b"].int)