+++ CHANGELOG.md
@@ -5,14 +5,17 @@
 - Experimental `listOfSpec` delegate.
 
 ### Changed
+- **API breaking** Config is deprecated, use `ObservableMeta` instead.
 - **API breaking** Descriptor no has a member property `defaultValue` instead of `defaultItem()` extension. It caches default value state on the first call. It is done because computing default on each call is too expensive.
 - Kotlin 1.5.10
 - Build tools 0.10.0
 - Relaxed type restriction on `MetaConverter`. Now nullables are available.
 
 ### Deprecated
+- Direct use of `Config`
 
 ### Removed
+- Public PluginManager mutability
 
 ### Fixed
 - Proper json array index treatment